+++ main.py
@@ -81,7 +81,7 @@ from dev.devices.sim_spectrometer import MySimSpectrometer
 
 
 # Define the application version
-version = "V3.0.3"
+version = "V3.1.0"
 
 # Set the size of the application's main window (width, height)
 size = (1200,700)
@@ -93,12 +93,11 @@ visible_notif_time = 3
 backup_directory = r"C:\Users\A068\Documents\Data\Default"
 #backup_directory = r"C:\Users\el-ei\Documents\Masterproject\Backup data"
 
-# Create an instance of a microscope setup
+# Create an instance of a microscope setup 28939
 # Remove 'Sim' from the class names to connect to real devices 
 gm_microscope = MyMicroscope("Maria Goeppert Mayer",
-    MyCamera("Camera", "11499", enable=True),
-    MySpectrometer("Spectrometer-VIS", "QEP06226", enable=True, dark_correction=True, integration_time=100),
-    MySpectrometer("Spectrometer-NIR", "NQ51B1981", integration_time=50),
+    MySpectrometer("Spectrometer", "xxxx", enable=True), #Need to implement Kymera spectrometer
+    MyCamera("Camera", "28939", enable=True),
     MyRotationMount("White-Light-Plzr", "COM6", "QEP06226", enable=False),
     MyRotationMount("Laser-Plzr", "COM4", "QEP06226", enable=False),
     MyShutter("White light - shutter", "26006167", enable=True, model="KST201", stage="FW103M"),
@@ -113,7 +112,8 @@ gm_microscope = MyMicroscope("Maria Goeppert Mayer",
     Avaspec 3648 spectrometer - Connexion is different from Oceanview spectrometers
 """
 lm_microscope = MyMicroscope("Lise Meitner",
-    MySpectrometer("Spectrometer", "xxxx", enable=True),
+    MySpectrometer("Spectrometer-VIS", "QEP06226", enable=True, dark_correction=True, integration_time=100),
+    MySpectrometer("Spectrometer-NIR", "NQ51B1981", integration_time=50),
     MyCamera("Camera", "11499", enable=True),
     MyShutter("White light - shutter", "26006167", enable=True, model="KST201", stage="FW103M"),
     MyShutter("Laser 750nm - shutter", "68801094", enable=False, model="KSC101"),